# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Overtime Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events, optionally filtered by roster
      parameters:
        - name: rosterId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Submit a new overtime event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get a single event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Reports ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Event:
      type: object
      required: [id, rosterId, date, description, defaultHours, dayType, multiplier, entries]
      properties:
        id:
          type: integer
        rosterId:
          type: integer
        date:
          type: string
          format: date
        description:
          type: string
        defaultHours:
          type: number
        dayType:
          type: string
        multiplier:
          type: number
          default: 1
        entries:
          type: array
          items:
            $ref: "#/components/schemas/EventEntry"
    EventEntry:
      type: object
      required: [id, employeeId, employeeName, offered, worked, hoursOffered, hoursAwarded]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        offered:
          type: boolean
        worked:
          type: boolean
        hoursOverride:
          type: ["number", "null"]
        hoursAwarded:
          type: number
        hoursOffered:
          type: number
    EventInput:
      type: object
      required: [rosterId, date, description, defaultHours, dayType, entries]
      properties:
        rosterId:
          type: integer
        date:
          type: string
          format: date
        description:
          type: string
          minLength: 1
        defaultHours:
          type: number
          minimum: 0
        dayType:
          type: string
        multiplier:
          type: number
          minimum: 0
          default: 1
        entries:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/EventEntryInput"
    EventEntryInput:
      type: object
      required: [employeeId, offered, worked]
      properties:
        employeeId:
          type: integer
        offered:
          type: boolean
        worked:
          type: boolean
        hoursOverride:
          type: ["number", "null"]

    # ── Reports ───────────────────────────────────────────────────────────────
...